+++ CHANGELOG.md
@@ -38,6 +38,9 @@
 
 * `transform/clr` component: Added the option to set the `axis` along which to apply CLR. Possible to override
   on workflow level as well (PR #767).
+  
+* `annotate/celltypist` component: Added a CellTypist annotation component (PR #825).
+
 
 * `workflows/annotation/harmony_knn`: workflow: Cell-type annotation based on harmony integration with KNN label transfer (PR #836).
 
@@ -65,6 +68,8 @@
 ## DOCUMENTATION
 
 * Update authorship of components (PR #835).
+
+* `resources_test_scripts/annotation_test_data.sh` script: Train and save a custom CellTypist model (PR #825).
 
 # openpipelines 1.0.0-rc6